refactor: Adjust content.ts to use SRP principle and functional programming
